TelegramRadar 2.0/requirements.txt
telethon>=1.34
rich>=13.7
python-dotenv>=1.0
aiohttp>=3.9
fpdf2>=2.7
